treewide: Update usage of compiler options.
